pyproject.toml
[tool.poetry]
name = "crac-server"
version = "0.1.0"
description = "Server for controlling the Dome and the Telescope"
authors = ["Alessio Caradossi <alessio@theeyes.org>"]
license = "GPL3"

[tool.poetry.dependencies]
python = "^3.10"
crac-protobuf = { path = "../crac-protobuf/dist/crac-protobuf-0.1.7.tar.gz" }
grpcio-tools = "^1.43.0"
gpiozero = "^1.6.2"
astropy = "^5.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
